requirements.txt.txt
pandas>=1.5.0
numpy>=1.23.0
python-Levenshtein>=0.21.0
homoglyphs>=2.2.0
requests>=2.28.0
tqdm>=4.64.0
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
datasets>=2.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0